Further trimmed down extension activation.
